+++ spaghetto/Parsing/Nodes/BinaryExpressionNode.cs
@@ -1,0 +1,65 @@
+﻿namespace spaghetto.Parsing.Nodes
+{
+    internal class BinaryExpressionNode : SyntaxNode
+    {
+        private SyntaxNode left;
+        private SyntaxToken operatorToken;
+        private SyntaxNode right;
+
+        public BinaryExpressionNode(SyntaxNode left, SyntaxToken operatorToken, SyntaxNode right)
+        {
+            this.left = left;
+            this.operatorToken = operatorToken;
+            this.right = right;
+        }
+
+        public override NodeType Type => NodeType.BinaryExpression;
+
+        public override SValue Evaluate(Scope scope)
+        {
+            var leftRes = left.Evaluate(scope);
+            var rightRes = right.Evaluate(scope);
+
+            switch (operatorToken.Type)
+            {
+                case SyntaxType.Plus:
+                    return leftRes.Add(rightRes);
+                case SyntaxType.Minus:
+                    return leftRes.Sub(rightRes);
+                case SyntaxType.Div:
+                    return leftRes.Div(rightRes);
+                case SyntaxType.Mul:
+                    return leftRes.Mul(rightRes);
+                case SyntaxType.Mod:
+                    return leftRes.Mod(rightRes);
+                case SyntaxType.EqualsEquals:
+                    return leftRes.Equals(rightRes);
+                case SyntaxType.Idx:
+                    return leftRes.Idx(rightRes);
+                case SyntaxType.LessThan:
+                    return leftRes.LessThan(rightRes);
+                case SyntaxType.LessThanEqu:
+                    return leftRes.LessThanEqu(rightRes);
+                case SyntaxType.GreaterThan:
+                    return leftRes.GreaterThan(rightRes);
+                case SyntaxType.GreaterThanEqu:
+                    return leftRes.GreaterThanEqu(rightRes);
+
+                default:
+                    throw new NotImplementedException();
+            }
+        }
+
+        public override IEnumerable<SyntaxNode> GetChildren()
+        {
+            yield return left;
+            yield return new TokenNode(operatorToken);
+            yield return right;
+        }
+
+        public override string ToString()
+        {
+            return "BinaryExprNode: op=" + operatorToken.Type;
+        }
+    }
+}
