+++ spaghetto/Parsing/Nodes/SyntaxNode.cs
@@ -1,0 +1,10 @@
+﻿namespace spaghetto.Parsing.Nodes
+{
+    public abstract class SyntaxNode
+    {
+        public abstract NodeType Type { get; }
+
+        public abstract SValue Evaluate(Scope scope);
+        public abstract IEnumerable<SyntaxNode> GetChildren();
+    }
+}
