+++ spaghetto/Parsing/Nodes/TokenNode.cs
@@ -1,0 +1,31 @@
+﻿namespace spaghetto.Parsing.Nodes
+{
+    /// <summary>
+    /// dummy node for tree view
+    /// </summary>
+    public class TokenNode : SyntaxNode
+    {
+        public override NodeType Type => NodeType.Token;
+        public SyntaxToken Token { get; set; }
+
+        public TokenNode(SyntaxToken token)
+        {
+            Token = token;
+        }
+
+        public override SValue Evaluate(Scope scope)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override IEnumerable<SyntaxNode> GetChildren()
+        {
+            return Enumerable.Empty<SyntaxNode>();
+        }
+
+        public override string ToString()
+        {
+            return "TokenNode: " + Token.Type.ToString() + " val=" + Token.Value?.ToString() + " text=" + Token.Text.ToString();
+        }
+    }
+}
