+++ spaghetto/Parsing/Parser.cs
@@ -8,8 +8,10 @@ using System.Reflection.Metadata;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading.Tasks;
+using spaghetto.Parsing.Nodes;
 
-namespace spaghetto {
+namespace spaghetto.Parsing
+{
     public class Parser {
         public List<SyntaxToken> Tokens { get; set; }
         public int Position = 0;
@@ -410,767 +412,6 @@ namespace spaghetto {
             }
 
             return left;
-        }
-    }
-
-    internal class NativeImportNode : SyntaxNode {
-        private SyntaxToken ident;
-
-        public NativeImportNode(SyntaxToken ident) {
-            this.ident = ident;
-        }
-
-        public override NodeType Type => NodeType.NativeImport;
-
-        public override SValue Evaluate(Scope scope) {
-            var val = scope.Get("nlimporter$$" + ident.Text);
-
-            if(val == null || val is not SNativeLibraryImporter importer) {
-                throw new Exception("Native library " + ident.Text + " not found!");
-            }
-
-            importer.Import(scope);
-            return SValue.Null;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren() {
-            yield return new TokenNode(ident);
-        }
-    }
-
-    internal class FunctionDefinitionNode : SyntaxNode {
-        private SyntaxToken? nameToken;
-        private List<SyntaxToken> args;
-        private SyntaxNode block;
-
-        public FunctionDefinitionNode(SyntaxToken? nameToken, List<SyntaxToken> args, SyntaxNode block) {
-            this.nameToken = nameToken;
-            this.args = args;
-            this.block = block;
-        }
-
-        public override NodeType Type => NodeType.FunctionDefinition;
-
-        public override SValue Evaluate(Scope scope) {
-            var f = new SFunction(scope, nameToken?.Text ?? "<anonymous>", args.Select((v) => v.Text).ToList(), block);
-            if(nameToken != null) scope.Set(nameToken.Value.Text, f);
-            return f;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren() {
-            if(nameToken != null) yield return new TokenNode(nameToken.Value);
-            foreach (var t in args) yield return new TokenNode(t);
-            yield return block;
-        }
-    }
-
-    internal class WhileNode : SyntaxNode {
-        private SyntaxNode condNode;
-        private SyntaxNode block;
-
-        public WhileNode(SyntaxNode condNode, SyntaxNode block) {
-            this.condNode = condNode;
-            this.block = block;
-        }
-
-        public override NodeType Type => NodeType.While;
-
-        public override SValue Evaluate(Scope scope) {
-            Scope whileScope = new(scope);
-            SValue lastVal = SValue.Null;
-
-            while (true) {
-                if (!condNode.Evaluate(whileScope).IsTruthy()) break;
-                var whileBlockRes = block.Evaluate(whileScope);
-                if (!whileBlockRes.IsNull()) lastVal = whileBlockRes;
-
-                if (whileScope.State == ScopeState.ShouldBreak) break;
-                if(whileScope.State != ScopeState.None) whileScope.SetState(ScopeState.None);
-            }
-
-            return lastVal;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren() {
-            yield return condNode;
-            yield return block;
-        }
-    }
-
-    internal class CastNode : SyntaxNode {
-        private SyntaxToken ident;
-        private SyntaxNode node;
-
-        public CastNode(SyntaxToken ident, SyntaxNode node) {
-            this.ident = ident;
-            this.node = node;
-        }
-
-        public override NodeType Type => NodeType.Cast;
-
-        public override SValue Evaluate(Scope scope) {
-            // TODO: maybe improve this
-            switch(ident.Text) {
-                case "int":
-                    return node.Evaluate(scope).CastToBuiltin(SBuiltinType.Int);
-                case "float":
-                    return node.Evaluate(scope).CastToBuiltin(SBuiltinType.Float);
-                case "string":
-                    return node.Evaluate(scope).CastToBuiltin(SBuiltinType.String);
-                case "list":
-                    return node.Evaluate(scope).CastToBuiltin(SBuiltinType.List);
-                default: throw new InvalidOperationException("INTERNAL: Cast was parsed successfully, but cast is not implemented for that!");
-            }
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren() {
-            yield return new TokenNode(ident);
-            yield return node;
-        }
-    }
-
-    internal class ForNode : SyntaxNode {
-        private SyntaxNode initialExpressionNode;
-        private SyntaxNode condNode;
-        private SyntaxNode stepNode;
-        private SyntaxNode block;
-
-        public ForNode(SyntaxNode initialExpressionNode, SyntaxNode condNode, SyntaxNode stepNode, SyntaxNode block) {
-            this.initialExpressionNode = initialExpressionNode;
-            this.condNode = condNode;
-            this.stepNode = stepNode;
-            this.block = block;
-        }
-
-        public override NodeType Type => NodeType.For;
-
-        public override SValue Evaluate(Scope scope) {
-            Scope forScope = new(scope);
-            SValue lastVal = SValue.Null;
-            initialExpressionNode.Evaluate(forScope);
-
-            while(true) {
-                if (!condNode.Evaluate(forScope).IsTruthy()) break;
-                var forBlockRes = block.Evaluate(forScope);
-                if (!forBlockRes.IsNull()) lastVal = forBlockRes;
-
-                if (forScope.State == ScopeState.ShouldBreak) break;
-                if(forScope.State != ScopeState.None) forScope.SetState(ScopeState.None);
-
-                stepNode.Evaluate(forScope);
-            }
-
-            return lastVal;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren() {
-            yield return initialExpressionNode;
-            yield return condNode;
-            yield return stepNode;
-            yield return block;
-        }
-    }
-
-    internal class IfNode : SyntaxNode {
-        public List<(SyntaxNode cond, SyntaxNode block)> Conditions { get; private set; } = new();
-
-        public override NodeType Type => NodeType.If;
-
-        public override SValue Evaluate(Scope scope) {
-            foreach((SyntaxNode cond, SyntaxNode block) in Conditions) {
-                var condRes = cond.Evaluate(scope);
-
-                if(condRes.IsTruthy()) {
-                    return block.Evaluate(new Scope(scope));
-                }
-            }
-
-            return SValue.Null;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren() {
-            foreach(var (cond, block) in Conditions) {
-                yield return cond;
-                yield return block;
-            }
-        }
-
-        internal void AddCase(SyntaxNode cond, SyntaxNode block) {
-            Conditions.Add((cond, block));
-        }
-
-        public override string ToString() {
-            return "IfNode:";
-        }
-    }
-
-    internal class ListNode : SyntaxNode {
-        private List<SyntaxNode> list;
-
-        public ListNode(List<SyntaxNode> list) {
-            this.list = list;
-        }
-
-        public override NodeType Type => NodeType.List;
-
-        public override SValue Evaluate(Scope scope) {
-            SList sList = new();
-
-            foreach(var n in list) {
-                sList.Value.Add(n.Evaluate(scope));
-            }
-
-            return sList;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren() {
-            foreach (var n in list) yield return n;
-        }
-
-        public override string ToString() {
-            return "ListNode:";
-        }
-    }
-
-    internal class IdentifierNode : SyntaxNode {
-        public SyntaxToken Token { get; private set; }
-
-        public IdentifierNode(SyntaxToken syntaxToken) {
-            this.Token = syntaxToken;
-        }
-
-        public override NodeType Type => NodeType.Identifier;
-
-        public override SValue Evaluate(Scope scope) {
-            return scope.Get((string)Token.Value) ?? SValue.Null;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren() {
-            yield return new TokenNode(Token);
-        }
-
-        public override string ToString() {
-            return "IdentNode:";
-        }
-    }
-
-    internal class IntLiteralNode : SyntaxNode {
-        private SyntaxToken syntaxToken;
-
-        public IntLiteralNode(SyntaxToken syntaxToken) {
-            this.syntaxToken = syntaxToken;
-        }
-
-        public override NodeType Type => NodeType.IntLiteral;
-
-        public override SValue Evaluate(Scope scope) {
-            var sint = new SInt((int)syntaxToken.Value);
-            return sint;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren() {
-            yield return new TokenNode(syntaxToken);
-        }
-
-        public override string ToString() {
-            return "IntLitNode:";
-        }
-    }
-
-    internal class FloatLiteralNode : SyntaxNode {
-        private SyntaxToken syntaxToken;
-
-        public FloatLiteralNode(SyntaxToken syntaxToken) {
-            this.syntaxToken = syntaxToken;
-        }
-
-        public override NodeType Type => NodeType.FloatLiteral;
-
-        public override SValue Evaluate(Scope scope) {
-            return new SFloat((float)syntaxToken.Value);
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren() {
-            yield return new TokenNode(syntaxToken);
-        }
-
-        public override string ToString() {
-            return "FloatLitNode:";
-        }
-    }
-
-    internal class StringLiteralNode : SyntaxNode {
-        private SyntaxToken syntaxToken;
-
-        public StringLiteralNode(SyntaxToken syntaxToken) {
-            this.syntaxToken = syntaxToken;
-        }
-
-        public override NodeType Type => NodeType.StringLiteral;
-
-        public override SValue Evaluate(Scope scope) {
-            return new SString((string)syntaxToken.Value);
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren() {
-            yield return new TokenNode(syntaxToken);
-        }
-
-        public override string ToString() {
-            return "StringLitNode:";
-        }
-    }
-
-    internal class CallNode : SyntaxNode {
-        public SyntaxNode ToCallNode { get; set; }
-        private List<SyntaxNode> argumentNodes;
-
-        public CallNode(SyntaxNode atomNode, List<SyntaxNode> argumentNodes) {
-            ToCallNode = atomNode;
-            this.argumentNodes = argumentNodes;
-        }
-
-        public override NodeType Type => NodeType.Call;
-
-        public override SValue Evaluate(Scope scope) {
-            var toCall = ToCallNode.Evaluate(scope);
-            var args = EvaluateArgs(scope);
-
-            return toCall.Call(scope, args);
-        }
-
-        public List<SValue> EvaluateArgs(Scope scope) {
-            var args = new List<SValue>();
-
-            foreach (var n in argumentNodes) args.Add(n.Evaluate(scope));
-            return args;
-        }
-            
-        public override IEnumerable<SyntaxNode> GetChildren() {
-            yield return ToCallNode;
-            foreach (var n in argumentNodes) yield return n;
-        }
-
-        public override string ToString() {
-            return "CallNode:";
-        }
-    }
-
-    internal class DotNode : SyntaxNode
-    {
-        public DotNode(SyntaxNode callNode)
-        {
-            CallNode = callNode;
-        }
-
-        public SyntaxNode CallNode { get; }
-        public List<SyntaxNode> NextNodes { get; internal set; } = new();
-
-        public override NodeType Type => NodeType.Dot;
-
-        public override SValue Evaluate(Scope scope)
-        {
-            var currentValue = CallNode.Evaluate(scope);
-
-            foreach(var node in NextNodes) {
-                if(node is IdentifierNode rvn) {
-                    var ident = rvn.Token;
-                    currentValue = currentValue.Dot(new SString((string)ident.Value));
-                }else if(node is CallNode cn) {
-                    if(cn.ToCallNode is IdentifierNode cnIdentNode) {
-                        var ident = cnIdentNode.Token;
-                        currentValue = currentValue.Dot(new SString((string)ident.Value)).Call(scope, cn.EvaluateArgs(scope));
-                    }else {
-                        throw new Exception("Tried to call a non identifier in dot not stack.");
-                    }
-                }else {
-                    throw new Exception("Unexpected node in dot node stack!");
-                }
-            }
-
-            return currentValue;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren()
-        {
-            yield return CallNode;
-
-            foreach (var node in NextNodes) yield return node;
-        }
-
-        public override string ToString() {
-            return "DotNode:";
-        }
-    }
-
-    internal class UnaryExpressionNode : SyntaxNode
-    {
-        private SyntaxToken token;
-        private SyntaxNode rhs;
-
-        public UnaryExpressionNode(SyntaxToken token, SyntaxNode rhs)
-        {
-            this.token = token;
-            this.rhs = rhs;
-        }
-
-        public override NodeType Type => NodeType.UnaryExpression;
-
-        public override SValue Evaluate(Scope scope)
-        {
-            switch(token.Type) {
-                case SyntaxType.Bang: return rhs.Evaluate(scope).Not();
-                case SyntaxType.Minus: return rhs.Evaluate(scope).ArithNot();
-                case SyntaxType.Plus: return rhs.Evaluate(scope);
-                default: throw new InvalidOperationException();
-            }
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren()
-        {
-            yield return new TokenNode(token);
-            yield return rhs;
-        }
-
-        public override string ToString() {
-            return "UnaryExpressionNode:";
-        }
-    }
-
-    internal class AssignVariableNode : SyntaxNode
-    {
-        private SyntaxToken ident;
-        private SyntaxNode expr;
-
-        public AssignVariableNode(SyntaxToken ident, SyntaxNode expr)
-        {
-            this.ident = ident;
-            this.expr = expr;
-        }
-
-        public override NodeType Type => NodeType.AssignVariable;
-
-        public override SValue Evaluate(Scope scope)
-        {
-            if(scope.Get(ident.Value.ToString()) == null) {
-                throw new InvalidOperationException("Can not assign to a non-existant identifier");
-            }
-
-            var val = expr.Evaluate(scope);
-            var key = ident.Value.ToString();
-            if (!scope.Update(key, val, out Exception ex)) throw ex;
-            return val;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren()
-        {
-            yield return new TokenNode(ident);
-            yield return expr;
-        }
-
-        public override string ToString() {
-            return "AssignVariableNode:";
-        }
-    }
-
-    internal class InitVariableNode : SyntaxNode
-    {
-        private SyntaxToken ident;
-        private SyntaxNode expr;
-        private readonly bool isFixedType = true;
-
-        public InitVariableNode(SyntaxToken ident, bool isFixedType)
-        {
-            this.ident = ident;
-            this.isFixedType = isFixedType;
-        }
-
-        public InitVariableNode(SyntaxToken ident, SyntaxNode expr, bool isFixedType)
-        {
-            this.ident = ident;
-            this.expr = expr;
-            this.isFixedType = isFixedType;
-        }
-
-        public override NodeType Type => NodeType.InitVariable;
-
-        public override SValue Evaluate(Scope scope)
-        {
-            if(scope.Get(ident.Value.ToString()) != null) {
-                throw new InvalidOperationException("Can not initiliaze the same variable twice!");
-            }
-
-            if (expr != null) {
-                var val = expr.Evaluate(scope);
-                val.TypeIsFixed = isFixedType;
-
-                scope.Set(ident.Value.ToString(), val);
-                return val;
-            }else {
-                if (isFixedType) throw new InvalidOperationException("Tried to initiliaze a fixed type variable with no value; this is not permitted. Use var% instead.");
-
-                scope.Set(ident.Value.ToString(), SValue.Null);
-                return SValue.Null;
-            }
-
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren()
-        {
-            yield return new TokenNode(ident);
-            if(expr != null) yield return expr;
-        }
-
-        public override string ToString() {
-            return "InitVariableNode:";
-        }
-    }
-
-    internal class BreakNode : SyntaxNode
-    {
-        public override NodeType Type => NodeType.Break;
-
-        public override SValue Evaluate(Scope scope)
-        {
-            scope.SetState(ScopeState.ShouldBreak);
-            return SValue.Null;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren()
-        {
-            return Enumerable.Empty<SyntaxNode>();
-        }
-
-        public override string ToString() {
-            return "BreakNode:";
-        }
-    }
-
-    internal class ContinueNode : SyntaxNode
-    {
-        public override NodeType Type => NodeType.Continue;
-
-        public override SValue Evaluate(Scope scope)
-        {
-            scope.SetState(ScopeState.ShouldContinue);
-            return SValue.Null;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren()
-        {
-            return Enumerable.Empty<SyntaxNode>();
-        }
-
-        public override string ToString() {
-            return "ContinueNode:";
-        }
-    }
-
-    internal class ReturnNode : SyntaxNode
-    {
-        public ReturnNode()
-        {
-        }
-
-        public ReturnNode(SyntaxNode returnValueNode)
-        {
-            ReturnValueNode = returnValueNode;
-        }
-
-        public SyntaxNode ReturnValueNode { get; }
-
-        public override NodeType Type => NodeType.Return;
-
-        public override SValue Evaluate(Scope scope)
-        {
-            scope.SetState(ScopeState.ShouldReturn);
-
-            if(ReturnValueNode != null) {
-                var v = ReturnValueNode.Evaluate(scope);
-                scope.SetReturnValue(v);
-            }
-
-            return SValue.Null;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren()
-        {
-            if (ReturnValueNode == null) yield break;
-            else yield return ReturnValueNode;
-        }
-
-        public override string ToString() {
-            return "ReturnNode:";
-        }
-    }
-
-    internal class BlockNode : SyntaxNode
-    {
-        private List<SyntaxNode> nodes;
-        private readonly bool createNewScope;
-
-        public BlockNode(List<SyntaxNode> nodes, bool createNewScope = true)
-        {
-            this.nodes = nodes;
-            this.createNewScope = createNewScope;
-        }
-
-        public override NodeType Type => NodeType.Block;
-
-        public override SValue Evaluate(Scope scope)
-        {
-            var lastVal = SValue.Null;
-            var blockScope = scope;
-            
-            if(createNewScope) blockScope = new Scope(scope);
-
-            foreach (var node in nodes) {
-                var res = node.Evaluate(blockScope);
-
-                if (!res.IsNull()) {
-                    lastVal = res;
-                }
-
-                if (scope.State == ScopeState.ShouldBreak
-                    || scope.State == ScopeState.ShouldContinue) return lastVal;
-
-                if (scope.State == ScopeState.ShouldReturn) {
-                    var v = scope.ReturnValue;
-                    return v;
-                }
-            }
-
-            return lastVal;
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren()
-        {
-            foreach(var node in nodes) yield return node;
-        }
-
-        public override string ToString() {
-            return "BlockNode:";
-        }
-    }
-
-    public abstract class SyntaxNode
-    {
-        public abstract NodeType Type { get; }
-
-        public abstract SValue Evaluate(Scope scope);
-        public abstract IEnumerable<SyntaxNode> GetChildren();
-    }
-
-    internal class BinaryExpressionNode : SyntaxNode
-    {
-        private SyntaxNode left;
-        private SyntaxToken operatorToken;
-        private SyntaxNode right;
-
-        public BinaryExpressionNode(SyntaxNode left, SyntaxToken operatorToken, SyntaxNode right)
-        {
-            this.left = left;
-            this.operatorToken = operatorToken;
-            this.right = right;
-        }
-
-        public override NodeType Type => NodeType.BinaryExpression;
-
-        public override SValue Evaluate(Scope scope)
-        {
-            var leftRes = left.Evaluate(scope);
-            var rightRes = right.Evaluate(scope);
-
-            switch(operatorToken.Type) {
-                case SyntaxType.Plus:
-                    return leftRes.Add(rightRes);
-                case SyntaxType.Minus:
-                    return leftRes.Sub(rightRes);
-                case SyntaxType.Div:
-                    return leftRes.Div(rightRes);
-                case SyntaxType.Mul:
-                    return leftRes.Mul(rightRes);
-                case SyntaxType.Mod:
-                    return leftRes.Mod(rightRes);
-                case SyntaxType.EqualsEquals:
-                    return leftRes.Equals(rightRes);
-                case SyntaxType.Idx:
-                    return leftRes.Idx(rightRes);
-                case SyntaxType.LessThan:
-                    return leftRes.LessThan(rightRes);
-                case SyntaxType.LessThanEqu:
-                    return leftRes.LessThanEqu(rightRes);
-                case SyntaxType.GreaterThan:
-                    return leftRes.GreaterThan(rightRes);
-                case SyntaxType.GreaterThanEqu:
-                    return leftRes.GreaterThanEqu(rightRes);
-
-                default:
-                    throw new NotImplementedException();
-            }
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren()
-        {
-            yield return left;
-            yield return new TokenNode(operatorToken);
-            yield return right;
-        }
-
-        public override string ToString() {
-            return "BinaryExprNode: op=" + operatorToken.Type;
-        }
-    }
-
-    public class BoolNode : SyntaxNode
-    {
-        public override NodeType Type => NodeType.BooleanLiteral;
-
-        public bool Value { get; set; }
-
-        public BoolNode(bool value)
-        {
-            Value = value;
-        }
-
-        public override SValue Evaluate(Scope scope)
-        {
-            return new SInt(Value ? 1 : 0);
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren()
-        {
-            return Enumerable.Empty<SyntaxNode>();
-        }
-
-        public override string ToString() {
-            return "BoolNode: " + Value;
-        }
-    }
-
-    // dummy node for tree view
-    public class TokenNode : SyntaxNode
-    {
-        public override NodeType Type => NodeType.Token;
-        public SyntaxToken Token { get; set; }
-
-        public TokenNode(SyntaxToken token)
-        {
-            Token = token;
-        }
-
-        public override SValue Evaluate(Scope scope)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override IEnumerable<SyntaxNode> GetChildren()
-        {
-            return Enumerable.Empty<SyntaxNode>();
-        }
-
-        public override string ToString() {
-            return "TokenNode: " + Token.Type.ToString() + " val=" + Token.Value?.ToString() + " text=" + Token.Text.ToString();
         }
     }
 }
