+++ spaghettoCLI/Program.cs
@@ -1,4 +1,5 @@
 ﻿using spaghetto;
+using spaghetto.Parsing.Nodes;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
 using System.Text;
